feat(header): extra header and tabbar container from container file
